+++ src/plugins/core/node.rs
@@ -163,9 +163,13 @@ impl NodePlugin {
             pr.set_message(format!("download {tarball_name}"));
             HTTP.download_file(url.clone(), local, Some(pr)).await?;
         }
-        if *env::MISE_NODE_VERIFY && !tv.checksums.contains_key(&tarball_name) {
-            tv.checksums
-                .insert(tarball_name, self.get_checksum(ctx, local, version).await?);
+        let platform_info = tv
+            .lock_platforms
+            .entry(self.get_platform_key())
+            .or_default();
+        platform_info.url = Some(url.to_string());
+        if *env::MISE_NODE_VERIFY && platform_info.checksum.is_none() {
+            platform_info.checksum = Some(self.get_checksum(ctx, local, version).await?);
         }
         self.verify_checksum(ctx, tv, local)?;
         Ok(())
@@ -599,7 +603,7 @@ fn arch(settings: &Settings) -> &str {
     let arch = settings.arch();
     if arch == "x86" {
         "x86"
-    } else if arch == "x86_64" {
+    } else if arch == "x64" {
         "x64"
     } else if arch == "arm" {
         if cfg!(target_feature = "v6") {